% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNH4-package.R
\docType{package}
\name{DNH4-package}
\alias{DNH4}
\alias{DNH4-package}
\title{DNH4: Crawling for Daum News Text}
\description{
Provides some utils to get Korean text sample from news articles in Daum which is popular news portal service in Korea.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://forkonlp.github.io/DNH4}
  \item \url{https://github.com/forkonlp/DNH4}
  \item Report bugs at \url{https://github.com/forkonlp/DNH4/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (\href{https://orcid.org/0000-0001-6474-2570}{ORCID})

}
\keyword{internal}
