% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComment.R
\name{getAllComment}
\alias{getAllComment}
\title{Get All Comment}
\usage{
getAllComment(turl, sort = c("RECOMMEND", "LATEST"))
}
\arguments{
\item{turl}{like 'http://v.media.daum.net/v/20161117210603961'.}

\item{sort}{you can select RECOMMEND, LATEST. RECOMMEND is Default.}
}
\value{
Get data.frame.
}
\description{
Get daum news comments
}
