% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComment.R
\name{getComment}
\alias{getComment}
\title{Get Comment}
\usage{
getComment(turl = url, limit = 10, parentId = 0, sort = c("RECOMMEND",
  "LATEST"))
}
\arguments{
\item{turl}{like 'http://v.media.daum.net/v/20161117210603961'.}

\item{limit}{is number of comment. defult is 10.}

\item{parentId}{defult is 0.}

\item{sort}{you can select RECOMMEND, LATEST. RECOMMEND is defult.}
}
\value{
Get data.frame.
}
\description{
Get daum news comments
if you want to get data only comment, enter commend like below.
getComment(url)$result$commentList[[1]]
}
