% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getContent}
\alias{getContent}
\title{Get Content}
\usage{
getContent(url = url)
}
\arguments{
\item{url}{is daum news link.}
}
\value{
Get data.frame(url,datetime,press,title,content).
}
\description{
Get daum news content from links.
}
