% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMainCategory.R
\name{getMainCategory}
\alias{getMainCategory}
\title{Get News Main Categories}
\usage{
getMainCategory(fresh = FALSE)
}
\arguments{
\item{fresh}{If TRUE, get data from internet. 
Default is FALSE which is return with cache.}
}
\value{
Get data.frame(chr:cate_name, chr:url).
}
\description{
Get daum news main category names and ids recently.
}
\examples{
  getMainCategory()

}
