% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaxPageNum.R
\name{getMaxPageNum}
\alias{getMaxPageNum}
\title{Get Max Page Number}
\usage{
getMaxPageNum(turl = url)
}
\arguments{
\item{turl}{is target url include breakingnews, category url, date without regDate like below.
'http://media.daum.net/breakingnews/politics/president?regDate=20161104'}
}
\value{
Get numeric
}
\description{
Get Max Page Number
}
