% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubCategory.R
\name{getSubCategory}
\alias{getSubCategory}
\title{Get News Sub Categories}
\usage{
getSubCategory(categoryUrl = "/society")
}
\arguments{
\item{categoryUrl}{Main category url in daum news. Only 1 value is passible. Default is society.}
}
\value{
Get data.frame(chr:sub_cate_name, chr:url).
}
\description{
Get daum news sub category names and urls recently.
}
