% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrlList.R
\name{getUrlList}
\alias{getUrlList}
\title{Get Url List}
\usage{
getUrlList(turl = url)
}
\arguments{
\item{turl}{is target url daum news.}
}
\value{
a [tibble][tibble::tibble-package](news_title, news_links).
}
\description{
Get daum news titles and links from target url.
}
