% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNH4-package.R
\docType{package}
\name{DNH4-package}
\alias{DNH4}
\alias{DNH4-package}
\title{DNH4: Crawling for daum news text}
\description{
Provides some utils to get korean text sample from news articles
            in daum which is popular news portal servise in korea.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/forkonlp/D2H4}
  \item Report bugs at \url{http://github.com/forkonlp/D2H4/issues}
}

}
\author{
\strong{Maintainer}: ChanYub Park \email{mrchypark@gmail.com}

}
\keyword{internal}
