% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComment.R
\name{getComment}
\alias{getComment}
\title{Get Comment}
\usage{
getComment(turl = url, limit = 10, parentId = 0,
  sort = c("RECOMMEND", "LATEST"), type = c("df", "list"))
}
\arguments{
\item{turl}{like 'http://v.media.daum.net/v/20161117210603961'.}

\item{limit}{is number of comment. defult is 10. You can set "all" to get all comments.}

\item{parentId}{defult is 0.}

\item{sort}{you can select RECOMMEND, LATEST. RECOMMEND is defult.}

\item{type}{return type. Defult is data.frame. It may sometimes warnning message.}
}
\value{
Get data.frame.
}
\description{
Get daum news comments
}
